feeds/requirements.txt
aiohttp
aiofiles
google-cloud-bigquery
google-cloud-storage
psycopg2
